fix(store): rename file useUserStore
